Adding our standardized settings files.
We are now going to be handling Drupal settings differently in our git
repositories. See
https://support.giantrabbit.com/projects/allgr/wiki/Configuring_DrupalCiviCRM_settings_files
for the details of this change, but the gist of it is that we're going to be
setting environment variables for each site environment that will allow us to
detect and set sensible defaults for the database settings and other
environment specific settings.  This allows us to commit the settings.php
file to the repository and keep sensitive info, like the database password,
not in the repo and in a settings.local.php file that's ignored by git.