Added flood_control module.
There appears to be some kind of vulnerability which the CalMen site was being probed for. In this case,
the scans were coming from a single IP address, which we blocked in .htaccess, but we are adding
the flood control mdoule to allow for more granular and more user-sided control of this kind of throttling.