Removing phone validation Issue #25653
CMG reported that community members trying to register for the first
time are required to enter the phone number in the following format:
XXX-XXX-XXXX.  If not in that format, the page refreshes an error
message about the phone format and without any of
their previously entered data.  The quickest solution is to simply
remove phone validation until it comes a problem for the CMG team.