Adding .well-known/.htaccess for Let's Encrypt.
We are using Let's Encrypt for SSL Certs. Our current process for obtaining
and installing certs uses Let's Encrypt's client, Certbot, and its webroot
plugin. The webroot plugin proves control of the domain by placing files
below the webroot in the .well-known directory. Drupal blocks access to any
directories that start with a . by default, so we have to add this .htaccess
file to the .well-known directory on our Drupal sites.